Update libsql-sqlite3 to SQLite 3.51.3 and regenerate the bundled libsql-ffi artifacts.

Restore libSQL compatibility across Makefile targets, legacy wasm configure flags, reserved-byte and encrypted replica behavior, vector VACUUM/integrity/FLOAT8 behavior, Tcl harness compatibility, WAL restart test cleanup, CRR reset ordering, and extension CI lockfile coverage.

+++ libsql-sqlite3/src/btmutex.c
@@ -185,7 +185,7 @@ int sqlite3BtreeHoldsMutex(Btree *p){
 */
 static void SQLITE_NOINLINE btreeEnterAll(sqlite3 *db){
   int i;
-  int skipOk = 1;
+  u8 skipOk = 1;
   Btree *p;
   assert( sqlite3_mutex_held(db->mutex) );
   for(i=0; i<db->nDb; i++){
